backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.9.0
pydantic-settings>=2.6.0
email-validator>=2.0.0
pdfminer.six==20231228
PyPDF2==3.0.1
python-docx==1.1.0
scikit-learn==1.6.1
google-generativeai>=0.8.0
reportlab==4.0.9
aiosmtplib==3.0.1
python-dotenv==1.0.1
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
